requirements.txt
aiofiles==22.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.20.0
altair==5.5.0
altgraph @ file:///AppleInternal/Library/BuildRoots/4~B7IEugDHS-fAL-bTRqAKCb_qmc84naJzmK7469Q/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
bottle==0.13.4
browser-cookie3==0.19.1
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.7
distro==1.9.0
exceptiongroup==1.3.1
f2==0.0.1.6
fastapi==0.128.0
ffmpeg-python==0.2.0
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.10.0
future @ file:///AppleInternal/Library/BuildRoots/4~B7IEugDHS-fAL-bTRqAKCb_qmc84naJzmK7469Q/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gitdb==4.0.12
GitPython==3.1.46
gmssl==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.11
imageio-ffmpeg==0.6.0
importlib-resources==6.1.0
importlib_metadata==8.7.1
iniconfig==2.1.0
iso8601==2.1.0
Jinja2==3.1.6
jiter==0.12.0
jsonpath-ng==1.6.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llvmlite==0.43.0
lz4==4.4.5
m3u8==3.6.0
macholib @ file:///AppleInternal/Library/BuildRoots/4~B7IEugDHS-fAL-bTRqAKCb_qmc84naJzmK7469Q/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
narwhals==2.15.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai==2.16.0
openai-whisper==20240930
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
ply==3.11
propcache==0.4.1
protobuf==4.23.0
proxy-tools==0.1.0
pyarrow==21.0.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyExecJS==1.5.1
Pygments==2.19.2
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pyobjc-core==12.0
pyobjc-framework-Cocoa==12.0
pyobjc-framework-Quartz==12.0
pyobjc-framework-Security==12.0
pyobjc-framework-UniformTypeIdentifiers==12.0
pyobjc-framework-WebKit==12.0
pypng==0.20220715.0
pytest==8.2.1
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
pywebview==6.1
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rich==13.7.1
rpds-py==0.27.1
six @ file:///AppleInternal/Library/BuildRoots/4~B7IEugDHS-fAL-bTRqAKCb_qmc84naJzmK7469Q/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
smmap==5.0.2
sniffio==1.3.1
starlette==0.49.3
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
tomli==2.4.0
torch==2.8.0
tornado==6.5.4
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.39.0
websockets==15.0.1
yarl==1.22.0
yt-dlp==2025.10.14
zipp==3.23.0
